Removed rdf_transformer2.py - Replaced rdf_transformer.py with it - Updated __init__.py and class names
